Force checking on Merch Name
