+++ SalesAccountBook/UI/Merch Screen/MerchIconView.swift
@@ -18,13 +18,7 @@ class MerchIconView: UIView {
             make.bottom.equalToSuperview().offset(-Constants.UI.Spacing.Small)
             make.width.equalTo(self.iconImage.snp.height)
         }
-        DispatchQueue.main.async {
-            self.iconImage.backgroundColor = Constants.UI.Color.Grey
-            self.iconImage.clipsToBounds = true
-            self.iconImage.layer.borderColor = UIColor.black.cgColor
-            self.iconImage.layer.borderWidth = 2.5
-            self.iconImage.layer.cornerRadius = self.iconImage.frame.width / 2
-        }
+        self.iconImage.backgroundColor = Constants.UI.Color.Grey
     }
     
     required init?(coder: NSCoder) {
